requirements.txt
mcp[cli]
mysql-connector-python
python-dotenv
